requirements.txt
openai
streamlit
langchain
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib